#2 Create resources for testing command module.
Here we wrote a little Python script, "cat.py", that precaiously simulate the
"cat" command. If some file is not found or cannot be opened, then it exits
with the return code 1 and prints a message.

It is, by itself, not very useful. However, we call it on tests.

Also, we created a file to be "cat'ed" by it.